# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nan Ayeyar API
  description: E-commerce API for rice trading platform
  version: 1.0.0
  contact:
    name: Nan Ayeyar Support
    email: support@nanayeyar.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.nanayeyar.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create order
      description: Create a new order (authenticated customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
                - shippingAddress
                - paymentType
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                        minimum: 1
                shippingAddress:
                  type: object
                  required:
                    - name
                    - address
                    - phone
                  properties:
                    name:
                      type: string
                    address:
                      type: string
                    phone:
                      type: string
                paymentType:
                  type: string
                  enum: [COD, ONLINE_TRANSFER]
                paymentMethodId:
                  type: string
                  description: Required for ONLINE_TRANSFER
                transactionId:
                  type: string
                  description: Required for ONLINE_TRANSFER
                paymentScreenshot:
                  type: string
                  description: Optional screenshot path for ONLINE_TRANSFER
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order created successfully
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Get order details (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  order:
                    $ref: '#/components/schemas/OrderDetail'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/orders/{id}/status:
    patch:
      tags:
        - Admin - Orders
      summary: Update order status
      description: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [PENDING, PROCESSING, ON_HOLD, SHIPPED, DELIVERED, CANCELED, RETURNED, REFUNDED]
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order status updated successfully
                  order:
                    $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Payment method endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        address:
          type: string
          example: 123 Main St, Yangon, Myanmar
        locale:
          type: string
          enum: [en, my]
          example: en
        role:
          type: string
          enum: [customer, admin]
          example: customer
        isEmailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        userId:
          type: string
          example: clr1234567890
        type:
          type: string
          enum: [AYA_BANK, KBZ_BANK, AYA_PAY, KBZ_PAY]
          example: AYA_BANK
        details:
          type: object
          example:
            accountName: "John Doe"
            accountNumber: "1234567890"
        createdAt:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        userId:
          type: string
          example: clr1234567890
        status:
          type: string
          enum: [PENDING, PROCESSING, ON_HOLD, SHIPPED, DELIVERED, CANCELED, RETURNED, REFUNDED]
          example: PENDING
        paymentType:
          type: string
          enum: [COD, ONLINE_TRANSFER]
          example: ONLINE_TRANSFER
        paymentMethodId:
          type: string
          example: clr1234567890
        totalAmount:
          type: number
          format: decimal
          example: 5000.00
        shippingAddress:
          type: object
          example:
            name: "John Doe"
            address: "123 Main St, Yangon, Myanmar"
            phone: "+959123456789"
        transactionId:
          type: string
          example: TXN123456789
        paymentScreenshot:
          type: string
          example: "/uploads/payment-screenshot.jpg"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            user:
              $ref: '#/components/schemas/User'
            paymentMethod:
              $ref: '#/components/schemas/PaymentMethod'
            refunds:
              type: array
              items:
                $ref: '#/components/schemas/Refund'
    OrderItem:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        orderId:
          type: string
          example: clr1234567890
        productId:
          type: string
          example: clr1234567890
        unitPrice:
          type: number
          format: decimal
          example: 2500.00
        quantity:
          type: integer
          example: 2
        metadata:
          type: object
          example:
            variety: "Jasmine"
            grade: "Premium"
            weight: "5kg"
    Refund:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        orderId:
          type: string
          example: clr1234567890
        amount:
          type: number
          format: decimal
          example: 5000.00
        refundedAt:
          type: string
          format: date-time
        reason:
          type: string
          example: "Customer requested refund"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string